Move forms into the view components themselves.
