--- src/components/forms/index.ts
@@ -1,8 +1,0 @@
-import LoginForm from "./LoginForm.vue";
-import EducatorSignUp from "./EducatorSignUp.vue";
-import StudentSignUp from "./StudentSignUp.vue";
-export {
-  LoginForm,
-  EducatorSignUp,
-  StudentSignUp
-};
