+++ src/views/EducatorAccountCreated.vue
@@ -1,0 +1,21 @@
+<template>
+  <v-container>
+    <v-row>
+      <v-col cols="12">
+        <h2>Account Created</h2>
+      </v-col>
+    </v-row>
+    <v-row>
+      <v-col cols="12">
+      <div>Thank you for creating your CosmicDS Educator Account</div>
+      <div>Please check your email for a confirmation link. After clicking the link in your email, you can log in and create your classroom codes.</div>
+      </v-col>
+    </v-row>
+    <v-spacer></v-spacer>
+    <v-row>
+      <v-col cols="12">
+        Already have an account? <a href="/login">Log in here</a>
+      </v-col>
+    </v-row>
+  </v-container>
+</template>
